requirements.txt
azure-identity
azure-mgmt-containerservice
azure-mgmt-monitor
schedule
time
msrestazure